requirements.txt
apex==0.9.10dev
binpacking==1.5.2
clip==0.2.0
datasets==2.14.4
decord==0.6.0
easydict==1.13
einops==0.8.1
flash_attn==2.8.2
h5py==3.14.0
huggingface_hub==0.31.2
huggingface_hub==0.34.3
jsonlines==4.0.0
matplotlib==3.10.8
modelscope==1.28.1
nltk==3.9.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
peft==0.17.0
Pillow==12.1.0
python-dotenv==1.2.1
qwen_vl_utils==0.0.14
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
spacy==3.8.11
svgwrite==1.4.3
timm==1.0.23
torch==2.7.1
torchcodec==0.9.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
